% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ddot_colors}
\alias{ddot_colors}
\title{DDOT color palette}
\format{
A named character vector with hex color codes.
}
\usage{
ddot_colors
}
\description{
An unnamed vector of hex codes for the DDOT colors.
}
\examples{
\dontrun{
lab_colors["ddot_blue"]
scale_color_manual(values = ddot_colors[c("ddot_gray", "ddot_blue", "ddot_red")])
}
}
\keyword{datasets}
