% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_address_mar.R
\name{geocode_address_mar}
\alias{geocode_address_mar}
\title{Geocode addresses using the MAR}
\usage{
geocode_address_mar(
  address = "955 L'Enfant Plaza SW",
  mar_api_key,
  cols = c("Latitude", "Longitude"),
  return_NA = TRUE
)
}
\arguments{
\item{address}{A character string with an address in Washington, DC}

\item{mar_api_key}{A character vector that identifies you to the MAR. Request a key at \url{https://developers.data.dc.gov/guide/getapikey}.}

\item{cols}{Selects which columns of the response you want to return. By default, returns latitude and longitude.}

\item{return_NA}{A logical indicating whether to return missing values}
}
\value{
Prints a data frame with information about the requested address. If \code{NULL}, prints "No results found."
}
\description{
This function queries the DC Master Address Repository (\url{https://octo.dc.gov/service/master-address-repository}) and returns details about the location.
You will need to download a (free) MAR API key and input it to the function (remember not to hard-code the key itself if you are pushing your code to GitHub.)\cr \cr
Warning: The code takes a long time to run. We recommend running once and then saving the result.
}
\examples{
# For a single address
\dontrun{
geocode_address("955 L'Enfant Plaza", mar_api_key = mar_api_key, cols = c("Latitude", "Longitude"))
}

# For batch geocoding
\dontrun{
sample_addresses |>
  mutate(geocode_result = map(address, ~ geocode_address_mar(.x, mar_api_key = mar_api_key))) |>
  unnest_wider(geocode_result)
}

}
