% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{lab_colors}
\alias{lab_colors}
\title{Lab color palette}
\format{
An unnamed character vector of hex color codes.
}
\usage{
lab_colors
}
\description{
An unnamed vector of hex codes for the Lab colors.
}
\examples{
lab_colors["lab_blue"]
scale_color_manual(values = lab_colors[c("lab_blue", "lab_red", "lab_light_blue")])

}
\keyword{datasets}
