% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{lab_colors_named}
\alias{lab_colors_named}
\title{Lab color palette}
\format{
A named character vector with hex color codes.
}
\usage{
lab_colors_named
}
\description{
A named vector of hex codes for the Lab colors.
}
\examples{
\dontrun{
lab_colors["lab_blue"]
scale_color_manual(values = lab_colors[c("lab_blue", "lab_red", "lab_light_blue")])
}
}
\keyword{datasets}
