% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ci_barplot.R
\name{make_ci_barplot}
\alias{make_ci_barplot}
\title{make_ci_barplot}
\usage{
make_ci_barplot(
  model = sample_model_1,
  treat_arms = c("sample_treatment1"),
  ci_level = 0.95,
  plot_caption = "Sample Plot Caption",
  ...
)
}
\arguments{
\item{model}{The regression whose results you want to plot}

\item{treat_arms}{A character vector for the treatment arms, as reported in the regression}

\item{ci_level}{The confidence level for the bar plot's intervals, expressed as a probability}

\item{plot_caption}{A character string with a caption for the plot}

\item{...}{Other arguments passed to \code{theme_lab()}}
}
\value{
A bar plot with a confidence interval on the treatment arms
}
\description{
Take regression output and make a bar plot which compares a control group with a treatment group, adding confidence intervals to the treatment groups.
}
