% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_sf_opendatadc}
\alias{read_sf_opendatadc}
\title{Download spatial data from Open Data DC}
\usage{
read_sf_opendatadc(
  url =
    "https://maps2.dcgis.dc.gov/dcgis/rest/services/DCGIS_DATA/Administrative_Other_Boundaries_WebMercator/MapServer/10",
  convert_crs = FALSE,
  print_map = TRUE
)
}
\arguments{
\item{convert_crs}{By default, the package reads in the shapefile in WGS 84 (ESPG 4326). If you plan to use the data for distance calculations, setting convert_crs = TRUE will convert the shapefile to DC GIS's preferred projection: The Maryland State Plane coordinate system of the American Datum of 1983 (NAD 83).}

\item{print_map}{Prints a simple shapefile of the data being read in as a side product.}
}
\value{
An sf object
}
\description{
Download and read in spatial data from a feature service on Open Data DC.
You can find a directory of spatial data at \url{https://maps2.dcgis.dc.gov/dcgis/rest/services/DCGIS_DATA}.
}
