% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_bbl}
\alias{sample_bbl}
\title{Sample Business License Data}
\format{
\subsection{\code{sample_bbl}}{

A data frame with 10,000 rows and 44 columns. The data is a random sample taken from data downloaded from OpenDataDC on August 25, 2025.
}
}
\source{
\url{https://opendata.dc.gov/datasets/85bf98d3915f412c8a4de706f2d13513_0/explore}
}
\usage{
sample_bbl
}
\description{
A sample of raw business license data pulled from Open Data
}
\keyword{datasets}
