% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lab.R
\name{theme_lab}
\alias{theme_lab}
\title{Make a plot with the Lab's theme}
\usage{
theme_lab(ppt = FALSE)
}
\arguments{
\item{ppt}{Logical; should the output be formatted for a Powerpoint slide? Defaults to FALSE}
}
\value{
A list:
\itemize{
\item if write = TRUE: path to saved ggplot object
\item if write = FALSE:ggplot object
}
}
\description{
Make a plot with the Lab's theme
}
