% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lab_seed.R
\name{gen_lab_seed}
\alias{gen_lab_seed}
\title{Generate Lab Random Seed}
\usage{
gen_lab_seed(seed_type = "date")
}
\arguments{
\item{seed_type}{A character string specifying \code{"date"} or \code{"sampled"}; defaults to \code{"date"}}
}
\value{
Prints the generated seed
}
\description{
Generate a seed for replicable work involving randomization or simulation. At The Lab, we use two types of seeds:
\itemize{
\item date seeds (public-relevant implementations)
\item sampled seeds (all other implementations)
}
We consider \emph{public-relevant} implementations to include situations like a random assignment of a program to some members of a waitlist, a random selection of some households to participate in a survey, and random assignment of hypothetical treatments during confirmatory randomization inference.\cr \cr
For either type of seed, run this function only once, and then include the seed in your code.
}
\examples{
gen_lab_seed()
gen_lab_seed("sampled")
}
