% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_lab}
\alias{theme_lab}
\title{Custom ggplot2 theme for The Lab}
\usage{
theme_lab(base_size = 12, palette = "lab", fonts_mayor = FALSE)
}
\arguments{
\item{base_size}{A numeric base font size.}

\item{palette}{A character string; one of \code{"lab"}, \code{"ddot"}, or a character vector of hex colors.}

\item{fonts_mayor}{A logical. If \code{TRUE}, uses the Neutra Text font family}
}
\value{
A list of ggplot2 theme and scale components.
}
\description{
Applies a minimal theme and custom discrete color/fill scales.
}
