% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_lab}
\alias{theme_lab}
\title{Custom ggplot2 theme for the Lab, modifiable}
\usage{
theme_lab(base_size = 12, palette = "lab", fonts_mmb = FALSE)
}
\arguments{
\item{base_size}{Base font size.}

\item{palette}{Either "lab", "ddot", or a character vector of hex colors.}

\item{fonts_mmb}{Logical. If TRUE, uses the Neutra Text font family}
}
\value{
A list of ggplot2 theme and scale components.
}
\description{
Applies a minimal theme and custom discrete color/fill scales.
}
